Extract scripting tests
