+++ test/test_scripting.py
@@ -1,0 +1,48 @@
+from __future__ import annotations
+
+import pytest
+import redis
+import redis.client
+
+from testtools import raw_command
+
+
+@pytest.mark.max_server('6.2.7')
+def test_script_exists(r):
+    # test response for no arguments by bypassing the py-redis command
+    # as it requires at least one argument
+    assert raw_command(r, "SCRIPT EXISTS") == []
+
+    # use single character characters for non-existing scripts, as those
+    # will never be equal to an actual sha1 hash digest
+    assert r.script_exists("a") == [0]
+    assert r.script_exists("a", "b", "c", "d", "e", "f") == [0, 0, 0, 0, 0, 0]
+
+    sha1_one = r.script_load("return 'a'")
+    assert r.script_exists(sha1_one) == [1]
+    assert r.script_exists(sha1_one, "a") == [1, 0]
+    assert r.script_exists("a", "b", "c", sha1_one, "e") == [0, 0, 0, 1, 0]
+
+    sha1_two = r.script_load("return 'b'")
+    assert r.script_exists(sha1_one, sha1_two) == [1, 1]
+    assert r.script_exists("a", sha1_one, "c", sha1_two, "e", "f") == [0, 1, 0, 1, 0, 0]
+
+
+@pytest.mark.parametrize("args", [("a",), tuple("abcdefghijklmn")])
+def test_script_flush_errors_with_args(r, args):
+    with pytest.raises(redis.ResponseError):
+        raw_command(r, "SCRIPT FLUSH %s" % " ".join(args))
+
+
+def test_script_flush(r):
+    # generate/load six unique scripts and store their sha1 hash values
+    sha1_values = [r.script_load("return '%s'" % char) for char in "abcdef"]
+
+    # assert the scripts all exist prior to flushing
+    assert r.script_exists(*sha1_values) == [1] * len(sha1_values)
+
+    # flush and assert OK response
+    assert r.script_flush() is True
+
+    # assert none of the scripts exists after flushing
+    assert r.script_exists(*sha1_values) == [0] * len(sha1_values)
